:art: Move Icon to components folder
